requirements-main.txt
flask
torch
Pillow
opencv-python 
gunicorn
RealESRGAN